Ensure transformation API works
